Added header component styling
