+++ src/components/header.js
@@ -1,0 +1,40 @@
+import { NavLink } from 'react-router-dom';
+
+const Header = () => {
+  const links = [
+    {
+      id: 1,
+      path: '/',
+      text: 'Home',
+    },
+    {
+      id: 2,
+      path: '/calculator',
+      text: 'Calculator',
+    },
+    {
+      id: 2,
+      path: '/quote',
+      text: 'Quote',
+    },
+  ];
+
+  return (
+    <header className="header">
+      <span>Math Magicians</span>
+      <nav>
+        <ul>
+          {links.map((link) => (
+            <li key={link.id}>
+              <NavLink to={link.path} activeClassName="active-link" exact>
+                {link.text}
+              </NavLink>
+            </li>
+          ))}
+        </ul>
+      </nav>
+    </header>
+  );
+};
+
+export default Header;
