--- src/components/navbar.js
@@ -1,37 +1,0 @@
-import { NavLink } from 'react-router-dom';
-
-const Navbar = () => {
-  const links = [
-    {
-      id: 1,
-      path: '/',
-      text: 'Home',
-    },
-    {
-      id: 2,
-      path: '/calculator',
-      text: 'About',
-    },
-    {
-      id: 2,
-      path: '/quote',
-      text: 'Quote',
-    },
-  ];
-
-  return (
-    <nav>
-      <ul>
-        {links.map((link) => (
-          <li key={link.id}>
-            <NavLink to={link.path} activeClassName="active-link" exact>
-              {link.text}
-            </NavLink>
-          </li>
-        ))}
-      </ul>
-    </nav>
-  );
-};
-
-export default Navbar;
